backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "voice-ai-platform"
version = "0.1.0"
description = "Real-Time Voice AI Assistant Platform"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "websockets==12.0",
    "httpx==0.27.0",

    # Config
    "pydantic-settings==2.2.1",
    "python-dotenv==1.0.1",

    # ASR
    "openai-whisper==20231117",
    "webrtcvad==2.0.10",
    "numpy==1.26.4",

    # LLM
    "openai==1.30.1",

    # TTS
    "TTS==0.22.0",
    "gTTS==2.5.1",
    "pydub==0.25.1",

    # Audio processing
    "librosa==0.10.2",
    "soundfile==0.12.1",
    "opuslib==3.0.1",

    # Storage
    "redis[hiredis]==5.0.4",
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "msgpack==1.0.8",
    "alembic==1.13.1",

    # Observability
    "structlog==24.1.0",

    # Utils
    "tenacity==8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.7",
    "pytest-cov==5.0.0",
    "fakeredis[aioredis]==2.23.2",
    "httpx==0.27.0",
    "locust==2.28.0",
    "ruff==0.4.4",
    "mypy==1.10.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "SIM"]
